converter-service/requirements.txt
aiodogstatsd==0.16.0.post0
aiohttp==3.8.6
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.2.0
circuitbreaker==2.0.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
dill==0.3.7
Django==4.2.5
django-cors-headers==4.2.0
django-filter==23.3
djangorestframework==3.14.0
exceptiongroup==1.1.3
flake8==6.1.0
flake8-html==0.4.3
flower==2.0.1
frozenlist==1.4.0
humanize==4.8.0
idna==3.4
imageio==2.31.6
imageio-ffmpeg==0.4.9
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.2
MarkupSafe==2.1.3
mccabe==0.7.0
molotov==2.6
moviepy==1.0.3
multidict==6.0.4
multiprocess==0.70.15
numpy==1.26.1
packaging==23.1
pika==1.3.2
Pillow==10.0.1
pluggy==1.3.0
proglog==0.1.10
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg==3.1.12
psycopg2-binary==2.9.9
pycodestyle==2.11.0
pyflakes==3.1.0
Pygments==2.16.1
PyJWT==2.8.0
pytest==7.4.2
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
salvo==0.2
six==1.16.0
sqlparse==0.4.4
stripe==7.1.0
tomli==2.0.1
tornado==6.3.3
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.8
yarl==1.9.2
